Rename from the default package name of react-native libraries
Using the com.reactlibrary package will cause conflcits if the user of this library uses another that's doing the same.

Can cause errors like

Error:Execution failed for task ':app:processDebugResources'.
:More than one library with package name 'com.reactlibrary'

Or

Error:Execution failed for task ':app:transformDexArchiveWithDexMergerForReleaseD8'
Program type already present: com.reactlibrary.BuildConfig
