+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactlibrary">
+          package="com.mkuczera">
     <uses-permission android:name="android.permission.VIBRATE"></uses-permission>
 </manifest>
-  
+
